feat: Add "@types/diacritics" npm dependency for diacritic removal
This commit adds the "@types/diacritics" npm dependency to the package.json file. The "@types/diacritics" library provides type definitions for the "diacritics" package, which is used to remove diacritics from strings. This dependency will be used in the "searchUserService" method of the "UsersService" class to preprocess search keys and filter users based on diacritic-insensitive matching.

#FT-789