refactor: Update `getDebtByTypeService` and `getDebtWhenDueService` in `DebtService`
This commit refactors the `getDebtByTypeService` and `getDebtWhenDueService` methods in the `DebtService` class of the `debt.service.ts` file.

In the `getDebtByTypeService` method, the code is modified to decrypt certain properties of the debts if they are encrypted. This ensures that the decrypted data is returned to the user.

In the `getDebtWhenDueService` method, the code is updated to fetch debts based on the provided `dueDate` parameter. Similar to the `getDebtByTypeService` method, the code also decrypts certain properties of the debts if they are encrypted.

These changes enhance the functionality of the `DebtService` by providing decrypted data to the user and improving the accuracy of debt retrieval based on due dates.🔧📝