refactor: Fix bug in CreateReplyCommentService method
The CreateReplyCommentService method in the CommentService was not correctly generating the _id for the new reply comment. This commit fixes the bug by using the Types.ObjectId() method from the mongoose library to generate a valid _id. The content of the reply comment is also now properly assigned. This bug fix ensures that new reply comments are created successfully.