feat: Add status property to Schedule schema
This commit adds a new property called "status" to the Schedule schema in the schedule.schema.ts file. The "status" property is of type string and is not required. It has a default value of "unRead" and can only have the values "unRead" or "readed". This enhancement allows for better tracking and management of the status of schedules.